move error Response to a model package
