+++ src/main/java/com/epam/microservice/Controller/TrainingWorkController.java
@@ -3,7 +3,6 @@ package com.epam.microservice.Controller;
 import com.epam.microservice.model.TrainingRequest;
 import com.epam.microservice.service.TrainingWorkService;
 import lombok.RequiredArgsConstructor;
-import org.slf4j.MDC;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
